LayoutTests/storage/indexeddb/deleteObjectStore-name-argument-required-expected.txt

Test IndexedDB deleteObjectStore required argument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
PASS db.deleteObjectStore(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
